// src/app/models/e-tables.ts
export enum ETables {

  employees = "Empleados",
  jobs = "Jobs",
  jobhistory = "Job Histories",
  departments = "Departamentos",
  charges = "Cargos",
  categories = "Categorias",
  shipping = "Compañias de Envios",
  contacts = "Contactos",
  details = "Detalles Pedidos",
  orders = "Pedidos",
  products = "Productos",
  suppliers = "Proveedores",
  concepts = "Conceptos",
  contracts = "Contratos",
  releases = "Libranzas",
  locations = "Lugares",
  payrolls = "Nómina",
  periods = "Periodos",
  type = "Tipos Contratos"

}